+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/calendar/CalendarPanel.java
@@ -65,8 +65,6 @@ import static javax.swing.BorderFactory.createTitledBorder;
 
 /**
  * A panel presenting a calendar for date/time selection.<br><br>
- * For a {@link CalendarPanel} without time fields use the {@link #dateCalendarPanel()} factory method.<br>
- * For a {@link CalendarPanel} with time fields use the {@link #dateTimeCalendarPanel()} factory method.<br><br>
  * Keyboard navigation:<br><br>
  * Previous/next year: CTRL + left/right arrow or down/up arrow.<br>
  * Previous/next month: SHIFT + left/right arrow or down/up arrow.<br>
@@ -74,6 +72,7 @@ import static javax.swing.BorderFactory.createTitledBorder;
  * Previous/next day: ALT + left/right arrow.<br>
  * Previous/next hour: SHIFT-ALT + left/right arrow or down/up arrow.<br>
  * Previous/next minute: CTRL-ALT + left/right arrow or down/up arrow.
+ * @see #builder()
  */
 public final class CalendarPanel extends JPanel {
 
@@ -128,9 +127,9 @@ public final class CalendarPanel extends JPanel {
   private final JLabel formattedDateLabel;
   private final boolean includeTime;
 
-  CalendarPanel(boolean includeTime) {
-    this.includeTime = includeTime;
-    LocalDateTime dateTime = LocalDateTime.now();
+  CalendarPanel(DefaultBuilder builder) {
+    this.includeTime = builder.includeTime;
+    LocalDateTime dateTime = builder.initialValue == null ? LocalDateTime.now() : builder.initialValue;
     yearValue = Value.value(dateTime.getYear(), dateTime.getYear());
     monthValue = Value.value(dateTime.getMonth(), dateTime.getMonth());
     dayValue = Value.value(dateTime.getDayOfMonth(), dateTime.getDayOfMonth());
@@ -228,17 +227,10 @@ public final class CalendarPanel extends JPanel {
   }
 
   /**
-   * @return a new {@link CalendarPanel} without time fields.
+   * @return a new {@link Builder} instance
    */
-  public static CalendarPanel dateCalendarPanel() {
-    return new CalendarPanel(false);
-  }
-
-  /**
-   * @return a new {@link CalendarPanel} with time fields.
-   */
-  public static CalendarPanel dateTimeCalendarPanel() {
-    return new CalendarPanel(true);
+  public static Builder builder() {
+    return new DefaultBuilder();
   }
 
   /**
@@ -246,6 +238,66 @@ public final class CalendarPanel extends JPanel {
    */
   public static Collection<Class<? extends Temporal>> supportedTypes() {
     return SUPPORTED_TYPES;
+  }
+
+  /**
+   * Builds a {@link CalendarPanel} instance.
+   */
+  public interface Builder {
+
+    /**
+     * Note that calling this method also sets {@link #includeTime(boolean)} to false
+     * @param initialValue the initial value
+     * @return this builder instance
+     */
+    Builder initialValue(LocalDate initialValue);
+
+    /**
+     * Note that calling this method also sets {@link #includeTime(boolean)} to true
+     * @param initialValue the initial value
+     * @return this builder instance
+     */
+    Builder initialValue(LocalDateTime initialValue);
+
+    /**
+     * @param includeTime if true then time fields are included (hours, minutes)
+     * @return this builder instance
+     */
+    Builder includeTime(boolean includeTime);
+
+    /**
+     * @return a new {@link CalendarPanel} based on this builder
+     */
+    CalendarPanel build();
+  }
+
+  private static final class DefaultBuilder implements Builder {
+
+    private LocalDateTime initialValue;
+    private boolean includeTime = false;
+
+    @Override
+    public Builder initialValue(LocalDate initialValue) {
+      this.initialValue = requireNonNull(initialValue).atStartOfDay();
+      return includeTime(false);
+    }
+
+    @Override
+    public Builder initialValue(LocalDateTime initialValue) {
+      this.initialValue = requireNonNull(initialValue);
+      return includeTime(true);
+    }
+
+    @Override
+    public Builder includeTime(boolean includeTime) {
+      this.includeTime = includeTime;
+      return this;
+    }
+
+    @Override
+    public CalendarPanel build() {
+      return new CalendarPanel(this);
+    }
   }
 
   void previousMonth() {
@@ -585,10 +637,10 @@ public final class CalendarPanel extends JPanel {
   private static JSpinner removeCtrlLeftRightArrowKeyEvents(JSpinner spinner) {
     InputMap inputMap = ((JSpinner.DefaultEditor) spinner.getEditor()).getTextField().getInputMap(WHEN_FOCUSED);
     //so it doesn't interfere with keyboard navigation when it has focus
-    inputMap.put(KeyStroke.getKeyStroke(VK_LEFT, CTRL_DOWN_MASK, false), "none");
-    inputMap.put(KeyStroke.getKeyStroke(VK_RIGHT, CTRL_DOWN_MASK, false), "none");
-    inputMap.put(KeyStroke.getKeyStroke(VK_LEFT, SHIFT_DOWN_MASK, false), "none");
-    inputMap.put(KeyStroke.getKeyStroke(VK_RIGHT, SHIFT_DOWN_MASK, false), "none");
+    inputMap.put(keyStroke(VK_LEFT, CTRL_DOWN_MASK), "none");
+    inputMap.put(keyStroke(VK_RIGHT, CTRL_DOWN_MASK), "none");
+    inputMap.put(keyStroke(VK_LEFT, SHIFT_DOWN_MASK), "none");
+    inputMap.put(keyStroke(VK_RIGHT, SHIFT_DOWN_MASK), "none");
 
     return spinner;
   }